Record reward for unsupervised env and fix a small bug.
The bug was that the health points could briefly be negative, and if
this happened at the end of an episode, it could result in the agent
receiving a higher penalty on death. We're not increasing the major
version of the package because the official Crafter report is not
released yet.